phase2/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pyjwt==2.9.0
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
email-validator==2.1.0
pydantic==2.10.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
httpx==0.26.0
bcrypt==4.1.2
